Removed white box test as it has been replaced.
Replaced by the a black box style test that is much easier to maintain and does not break whenever the implementation changes.
