+++ nh_clinical/tests/nh_clinical_staff_reallocation/__init__.py
@@ -8,4 +8,4 @@ from . import test_reallocate
 from . import test_staff_reallocation_complete
 from . import test_staff_reallocation_integration
 from . import test_staff_reallocation_multi_ward
-from . import test_default_users_black_box
+from . import test_default_users
